+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 # Change Log for React-id-swiper
+## [v1.6.4](https://github.com/kidjp85/react-id-swiper/compare/1.6.4...1.6.3) - Released on May 18th, 2018
+  * Provide lightweight version which drops those features below
+    - Virtual
+    - Keyboard
+    - Mouse wheel
+    - Zoom
+    - Lazy load image
+    - A11y
+    - History
+    - Hash-navigation
+    - Effect-cube
+    - Effect-flip
+    - Effect-coverflow
+
 
 ## [v1.6.3](https://github.com/kidjp85/react-id-swiper/compare/1.6.3...1.6.2) - Released on May 3rd, 2018
   * Upgrade swiper@4.2.6
@@ -24,16 +38,18 @@
 ## [v1.6.2](https://github.com/kidjp85/react-id-swiper/compare/1.6.2...1.6.1) - Released on April 1st, 2018
   * Upgrade swiper@4.2.2
   * Update unit test
+
 ## [v1.6.1](https://github.com/kidjp85/react-id-swiper/compare/1.6.1...1.5.8) - Released on February 17th, 2018
   * Upgrade swiper@4.1.6
   * Add unit test
+  
 ## [v1.5.8](https://github.com/kidjp85/react-id-swiper/compare/1.5.8...1.5.7) - Released on January 24th, 2018
   * Upgrade swiper@4.1.0
   * Fix bug for UMD build
 
 ## [v1.5.7](https://github.com/kidjp85/react-id-swiper/compare/1.5.7...1.5.6) - Released on December 20th, 2017
   * Add umd version
-  * Add new params 
+  * Add new params
     - ContainerEl
     - WrapperEl
     - renderCustomPrevButton
@@ -45,7 +61,7 @@
 ## [v1.5.6](https://github.com/kidjp85/react-id-swiper/compare/1.5.6...1.5.5) - Released on December 1st, 2017
   * Upgrade Swiper@4.0.7
   * Add jest. Expose rebuildSwiper instance method [PR 108](https://github.com/kidjp85/react-id-swiper/pull/108) by [timkindberg](https://github.com/timkindberg)
-  
+
 ## [v1.5.5](https://github.com/kidjp85/react-id-swiper/compare/1.5.5...1.5.4) - Released on November 15, 2017
   * Upgrade Swiper@4.0.6
   * Add support for parallax